user 代码优化

+++ src/main/java/com/onionch/webapp/website/controller/UserController.java
@@ -1,18 +1,16 @@
 package com.onionch.webapp.website.controller;
 
+import com.onionch.webapp.website.bean.enums.ResultCode;
+import com.onionch.webapp.website.bean.request.UserRequest;
 import com.onionch.webapp.website.bean.response.RestResponse;
-import com.onionch.webapp.website.bean.User;
 import com.onionch.webapp.website.service.UserService;
-import com.onionch.webapp.website.util.DESUtil;
-import org.apache.commons.lang.RandomStringUtils;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
-@RequestMapping("/api/v1/user")
+@RequestMapping("/api/v1/*")
 public class UserController {
 
     private static final Logger logger = Logger.getLogger(UserController.class);
@@ -20,36 +18,37 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    @RequestMapping(method = RequestMethod.GET)
-    private @ResponseBody RestResponse listAllUsers(){
-        List list=userService.listAll();
-        return RestResponse.response(200,"success",list);
+    @RequestMapping(value = "/users",method = RequestMethod.GET)
+    private @ResponseBody RestResponse listUsers(@RequestParam(value = "userName", required = false) String userName,
+                                                 @RequestParam(value = "pageIndex", required = false, defaultValue = "1") Integer pageIndex,
+                                                 @RequestParam(value = "pageSize", required = false, defaultValue = "50") Integer pageSize){
+        return userService.listAll(userName,pageIndex,pageSize);
     }
 
-    @RequestMapping(method = RequestMethod.POST)
-    private @ResponseBody RestResponse create(@RequestBody User user){
-        user.setuId(RandomStringUtils.randomNumeric(10));
-        String passwd=user.getUserPassword();
-        user.setUserPassword(DESUtil.encryptString(passwd));
-        userService.create(user);
-        return RestResponse.response(200,"success","");
+    @RequestMapping(value = "/users",method = RequestMethod.POST)
+    private @ResponseBody RestResponse create(@RequestBody UserRequest userRequest){
+        if (StringUtils.isEmpty(userRequest.getUserName())) {
+            return RestResponse.failure(ResultCode.PARAM_IS_BLANK, "username is empty");
+        }
+        if (StringUtils.isEmpty(userRequest.getUserPassword())) {
+            return RestResponse.failure(ResultCode.PARAM_IS_BLANK, "password is empty");
+        }
+        return userService.create(userRequest);
     }
 
-    @RequestMapping(value = "/{uId}",method = RequestMethod.DELETE)
+    @RequestMapping(value = "/users/{uId}",method = RequestMethod.DELETE)
     private @ResponseBody RestResponse delete(@PathVariable String uId){
-        userService.delete(uId);
-        return RestResponse.response(200,"success","");
+        if (StringUtils.isEmpty(uId)) {
+            return RestResponse.failure(ResultCode.PARAM_IS_BLANK, "uId is empty");
+        }
+        return userService.delete(uId);
     }
 
-    @RequestMapping(value = "/{uId}",method = RequestMethod.GET)
-    private @ResponseBody RestResponse getById(@PathVariable String uId){
-        User user=userService.selectByUid(uId);
-        return RestResponse.response(200,"success",user);
-    }
-
-    @RequestMapping(value = "/{uId}",method = RequestMethod.PUT)
-    private @ResponseBody RestResponse update(@PathVariable String uId,@RequestBody User user){
-        userService.update(user);
-        return RestResponse.response(200,"success","");
+    @RequestMapping(value = "/users/{uId}",method = RequestMethod.PUT)
+    private @ResponseBody RestResponse update(@PathVariable String uId,@RequestBody UserRequest userRequest){
+        if (StringUtils.isEmpty(uId)) {
+            return RestResponse.failure(ResultCode.PARAM_IS_BLANK, "uId is empty");
+        }
+        return userService.update(uId,userRequest);
     }
 }
